+++ include/simpledbus/base/Connection.h
@@ -8,16 +8,12 @@ namespace SimpleDBus {
 class Message;
 
 class Connection {
-  private:
-    ::DBusBusType _dbus_bus_type;
-    ::DBusConnection* _conn;
-    ::DBusError _err;
-
   public:
     Connection(::DBusBusType dbus_bus_type);
     ~Connection();
 
     void init();
+    void uninit();
 
     void add_match(std::string rule);
     void remove_match(std::string rule);
@@ -27,6 +23,12 @@ class Connection {
 
     void send(Message& msg);
     Message send_with_reply_and_block(Message& msg);
+
+  private:
+    bool _initialized = false;
+
+    ::DBusBusType _dbus_bus_type;
+    ::DBusConnection* _conn;
 };
 
 }  // namespace SimpleDBus